samples/misc/opencv_samples/mqtt_cam/requirements.txt
black>=19.3b0
paho-mqtt>=1.4.0
Pillow>=7.2.0
imutils>=0.5.3
PyYAML>=5.1.2
streamlit>=0.62.1
isort>=5.0.3
opencv-python>=4.1.1.26 # may fail on rpi